💄 버튼과 관련된 상수 추가
